Split the partials to their own files.
